+++ test/common/http/http2/codec_impl_test_util.h
@@ -38,12 +38,14 @@ private:
 
 class TestServerConnectionImpl : public ServerConnectionImpl, public TestCodecSettingsProvider {
 public:
-  TestServerConnectionImpl(Network::Connection& connection, ServerConnectionCallbacks& callbacks,
-                           Stats::Scope& scope,
-                           const envoy::config::core::v3::Http2ProtocolOptions& http2_options,
-                           uint32_t max_request_headers_kb, uint32_t max_request_headers_count)
+  TestServerConnectionImpl(
+      Network::Connection& connection, ServerConnectionCallbacks& callbacks, Stats::Scope& scope,
+      const envoy::config::core::v3::Http2ProtocolOptions& http2_options,
+      uint32_t max_request_headers_kb, uint32_t max_request_headers_count,
+      envoy::config::core::v3::HttpProtocolOptions::HeadersWithUnderscoresAction
+          headers_with_underscores_action)
       : ServerConnectionImpl(connection, callbacks, scope, http2_options, max_request_headers_kb,
-                             max_request_headers_count) {}
+                             max_request_headers_count, headers_with_underscores_action) {}
   nghttp2_session* session() { return session_; }
   using ServerConnectionImpl::getStream;